refactor(storage): share shouldBeUnsigned and TYPED_ARRAY_CTORS
Two more copy-pasted storage helpers:
- shouldBeUnsigned (number/integer with minimum >= 0) was duplicated in
  PostgresTabularStorage and SupabaseTabularStorage; lift it to
  BaseSqlTabularStorage, which already owns getNonNullType.
- TYPED_ARRAY_CTORS (TypedArray-name -> constructor map for decoding
  stored vectors) was duplicated in the postgres and sqlite tabular
  backends; export one copy from @workglow/storage's sql/ module.

Behavior-preserving.
