+++ packages/storage/src/sql/index.ts
@@ -10,3 +10,4 @@ export * from "./Dialect";
 export * from "./mapPostgresType";
 export * from "./PredicateBuilder";
 export * from "./PrefixDdl";
+export * from "./typedArrayCtors";
